Move M355 caselight to cpp
